requirements.txt
ccxt>=3.0.0
pandas>=1.5.0
numpy>=1.23.0
torch>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas-ta>=0.3.0
optuna>=3.0.0
scipy>=1.9.0
python-dotenv>=1.0.0 